유저 level 업데이트 API 명세

+++ src/main/java/cat/cat/helper/dto/FindCatHelperResponse.java
@@ -8,7 +8,7 @@ import lombok.Getter;
 @AllArgsConstructor
 public class FindCatHelperResponse {
     private Long helperId;
-    private Long exp;
+    private Double exp;
     private Long level;
     private boolean active;
